Encapsulate audio manager
