+++ src/lib.rs
@@ -1,4 +1,5 @@
 pub mod assets;
+pub mod audio;
 pub mod config;
 pub mod errors;
 pub mod tja;
